requirements.txt
Django==4.1.8
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
drf_yasg==1.21.5
rest_framework_simplejwt==0.0.2

asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
docopt==0.6.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pipreqs==0.4.11
PyJWT==2.6.0
pymongo==3.12.1
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
sqlparse==0.2.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
yarg==0.1.9